+++ src/ValueObject/IdList.php
@@ -7,9 +7,11 @@ namespace DigitalCraftsman\Ids\ValueObject;
 use DigitalCraftsman\Ids\ValueObject\Exception\DuplicateIds;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdAlreadyInList;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdClassNotHandledInList;
+use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainEveryId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainEveryId;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainId;
+use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesContainNoneIds;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListDoesNotContainSomeIds;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListIsNotEmpty;
 use DigitalCraftsman\Ids\ValueObject\Exception\IdListsMustBeEqual;
@@ -121,12 +123,7 @@ abstract class IdList implements \IteratorAggregate, \Countable
     /** @param static $idList */
     public function addIds(self $idList): static
     {
-        // TODO: Use must contain none ids
-        foreach ($idList as $id) {
-            if ($this->containsId($id)) {
-                throw new IdAlreadyInList($id);
-            }
-        }
+        $this->mustContainNoneIds($idList);
 
         $newIds = $this->ids;
         foreach ($idList as $id) {
@@ -186,9 +183,7 @@ abstract class IdList implements \IteratorAggregate, \Countable
     /** @param static $idList */
     public function removeIds(self $idList): static
     {
-        foreach ($idList as $id) {
-            $this->mustContainId($id);
-        }
+        $this->mustContainEveryId($idList);
 
         $ids = [];
         foreach ($this->ids as $id) {
@@ -343,6 +338,18 @@ abstract class IdList implements \IteratorAggregate, \Countable
         return true;
     }
 
+    public function notContainsEveryId(self $idList): bool
+    {
+        foreach ($idList as $id) {
+            if ($this->notContainsId($id)) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /** Opposite function is not called notContainsSomeIds, but containsNoneIds */
     public function containsSomeIds(self $idList): bool
     {
         foreach ($idList as $id) {
@@ -352,6 +359,17 @@ abstract class IdList implements \IteratorAggregate, \Countable
         }
 
         return false;
+    }
+
+    public function containsNoneIds(self $idList): bool
+    {
+        foreach ($idList as $id) {
+            if ($this->containsId($id)) {
+                return false;
+            }
+        }
+
+        return true;
     }
 
     /** @param static $idList */
@@ -453,11 +471,27 @@ abstract class IdList implements \IteratorAggregate, \Countable
         }
     }
 
+    /** @throws IdListDoesContainEveryId */
+    public function mustNotContainEveryId(self $idList): void
+    {
+        if (!$this->notContainsEveryId($idList)) {
+            throw new IdListDoesContainEveryId();
+        }
+    }
+
     /** @throws IdListDoesNotContainSomeIds */
     public function mustContainSomeIds(self $idList): void
     {
         if (!$this->containsSomeIds($idList)) {
             throw new IdListDoesNotContainSomeIds();
+        }
+    }
+
+    /** @throws IdListDoesContainNoneIds */
+    public function mustContainNoneIds(self $idList): void
+    {
+        if (!$this->containsNoneIds($idList)) {
+            throw new IdListDoesContainNoneIds();
         }
     }